Hotfix: 세종캠퍼스 회원가입 시 본전공 선택 & 마이보드 세종캠퍼스, 학번 매핑 수정

+++ src/pages/signUp/NewSignUpPage1.tsx
@@ -46,7 +46,7 @@ export function SignUp1Page() {
 
       const { koreapasUUID, nickname, studentId, firstMajorCode, firstMajorCampus } = res.data.data;
 
-      if (!firstMajorCode) {
+      if (firstMajorCampus === 'A' && !firstMajorCode) {
         alert('고파스의 ‘꽈톡’ 페이지에 접속 후 다시 회원가입을 시도해주세요.');
         sessionStorage.clear();
         navigate('/');
